Removed provider and old models
